+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # vend
 
-## Unreleased
+## 0.3.1 (2025-09-03)
 
 #### Added